# Atom prototype benchmark