--- a/opencog/benchmark/CMakeLists.txt
+++ b/opencog/benchmark/CMakeLists.txt
@@ -53,3 +53,15 @@ IF (HAVE_GUILE)
 		dl
 	)
 ENDIF (HAVE_GUILE)
+
+ADD_EXECUTABLE (atom_bench
+	PagedAtoms.cc
+	atom_bench.cc
+)
+
+TARGET_LINK_LIBRARIES (atom_bench
+	atombase
+	atomspace
+	${ATOMSPACE_LIBRARIES}
+	${COGUTIL_LIBRARY}
+)