Added an experimental benchmark to prototype and time an alternate paged-based storage for atoms in order to determine the potential speed improvements to be gained by improving locality of reference for atoms by storing them in adjacent memory.