@linas 2017-07-25
Sounds interesting.  There are at least five different types of locality problems:
1) data aliases to the same cache lines, and thus cache lines get ejected too often.
2) data sits on many many 4K pages, thus overwhelms the number of available TLB entries.  The linux hugepage support can help with this.
3) data sits in different pages in RAM, causing inefficient swapin/swapout behavior.
4) overuse of pointers, pointer-chasing when in-place indexes in contiguous arrays would be better.
5) Cache lines filled with useless junk, wasted space, instead of being dense with needed data.

I don't know that we can do anything about 1 and solving 3 seems pointless.  That leaves 2 & 4,5.

For 2, a quickie experiment might be to use linux hugepages for opencog. My datasets are dozens of GBytes or more, so TLB thrashing is presumably extremely common.  So, experiments include:
a) does linux hugepage help with performance?
b) you mentioned some sort of whizzy malloc/free library on the mailing list ... it sounded cool .. have you tried this? What happened?

For 4 and 5, I did try to use `std::emplace_back` and similar ctors when building the outgoing set of an atom. The idea was that all of the outgoing set would be in an array, on the same cache line. This made no difference for small datasets. Did  not try for big ones.

My typical data is this:
```
EvaluationLink
      PredicateNode "pair"
      ListLink
            WordNode "foo"
            WordNode "bar"
``` 
It would be nice if we could get both the ListLink and the EvaluationLink located on the same cache line. Note, however: what I need to repeatedly access are the various values hanging on the EvaluationLink: the CountTruthValue, the FloatValue holding the frequency (probability) the FloatValue holding the mutual information, the FloatValue holding the cosine angle.   So, its not so much the locality of the atoms that matters, but the locality of the values attached to the atoms.

The problem with locality is "wasted airspace", Suppose you did manage to put both links, and all four values (count, freq, MI, cosine) on the same cache line. My algos usually examine ONLY the mi, or ONLY the cosine, so putting all these on the same cacheline is a waste: it just uses dead space.

It would be much much better if all   CountTruthValue's were on the same hugepage, if all cosines were on the same hugepage.... if we could alloc atom values based in their key. !! That might be awesome..

Also: need to ask "what is an atom, really?"  In my datasets, I see about 2KBytes RAM per atom. Given that `sizeof(Atom)` is maybe 150 or 200 bytes, where's the rest?  The answer seems to be:

* 100-200 bytes in outgoing set
* 500+ bytes in the incoming set
* 500+bytes in the various values attached to atom.
* 400 bytes in the atomspace indexes.

or something like that.  Not sure.  What this means is that a typical atom is actually 10x bigger than the `sizeof (class Atom)`, and that typical acces to one atom touches many, or most, or maybe *all* of these different places in RAM.   I don't know ho to even measure the non-locality of this access, and its really hard to guess which of these are important, and which are not.