@linas 2017-07-25
So, I'm skimming your code, and so perhaps I totally misunderstand it, but ... it seems you are trying to store the string-name of a node in contiguous memory with the node itself.  This seems to me like not just a waste, but a performance-deaccelerator, since it packs useless data (the node name) with the node, thus wasting cache-line space!

The point is that there are almost never any accesses to the node-name, its almost a set-and-forget value. What does happen a lot are atomspace queries, to obtain a node of a given name and type. That works by creating an atom, computing the hash for it, and throwing away the atom.  There is a content compare only on hash collisions.  All of the rest of the time, the atoms are always access by their handles, and their handles are used mostly to find values on them - truth values or whatever, and that is where the CPU time all goes. Trying to drag along the string name with the node seems to mostly clutter RAM with boring data that is never accessed. See what i mean?