@inflector 2017-07-26
In general, I have found that custom allocators for common and high-volume C++ objects is often much much faster than the built-ins. This appears to be true for Atoms as well.

I just added some more details from my notes and some benchmark results to the ReadMe, as well as more comments to explain the code. Also note the diagrams in the ReadMe which show the current design memory layout. I also performed a fairly detailed analysis of the current Atom memory requirements which can be found in that document.

> seems you are trying to store the string-name of a node in contiguous memory with the node itself

The code stores the std::string object there but std::string doesn't store the actual string as that is implemented via a malloc or equivalent somewhere else, std::string is a fixed 8-bytes.

> It would be much much better if all CountTruthValue's were on the same hugepage, if all cosines were on the same hugepage.... if we could alloc atom values based in their key. !! That might be awesome..

The prototype did not implement any values storage but what you outline above, i.e. allocating atom values based on their key, would be fairly straightforward. If we allocated values out of memory-pages ordered by key, then atom, then the values for a given predicate would be contiguous. This should exhibit similar speed improvements as indicated in the test notes. 

The other possibility, and one that would be higher performance, and less memory in most use scenarios, would be to have separate pages for each different key, that way each page would contain closely-packed typed values for the key. 

This would be a bit more work as you'd have to maintain a directory of pages for each different key, but it would save you 64-bits per entry. If the pages were fairly small, like 1K each, this might not be too much overhead and the number of values stored per page would increase by 50% for this approach, assuming 8-byte Atom pointers and 8-bytes for value data.