@inflector 2017-07-26
The images are not visible until this Pull Request is merged as they are MarkDown references. So I am adding them here so you can seem them:

![image](http://i.imgur.com/lxy1o4D.png)

and

![image](http://i.imgur.com/8o023sj.png)