@inflector 2017-07-26
Time to traverse 1 million EvaluationLinks looking for a match on a particular node:

```
2.54s - current AtomSpace
0.11s - paged-Atoms prototype
```

of this time, for our current AtomSpace, 1.05s is spent in get_handles_by_type building the HandleSeq list.