@linas 2017-07-26
> I have found that custom allocators for common and high-volume C++ objects is often much much faster than the built-ins. This appears to be true for Atoms as well.

Then perhaps we should switch to one of these.  

> std::string is a fixed 8-bytes

Yeah. Since the name and the outgoing set are immutable, it makes sense to ditch `std::string` for storage and just use fixed-length arrays for this, for the outgoing set, and for the lists of values.  The only real problem with this suggestion is that our current API returns `const std::string&` which makes such a change ugly/unworkable :-(

> separate pages for each different key
Separate malloc pools for each key, yes, exactly.