@inflector 2017-07-27
Linas wrote:
> I'm concerned that you are optimizing for a situation where you have tiny incoming sets. I would like you to do your prototyping and measuring on some of my actual datasets.

and:
> Your diagram shows "set-by-type" -- this was added about 1/2 year ago, maybe more recently, and gave an instant 10x performance improvement in incoming set traversal. This was an important change, yet its not visible in your altered diagram.

The code in this prototype is designed to test the storage for atoms that uses B*-tree like storage. It is like a b-tree except the pages are in-memory and not paged to disk.  The first phase of this test indicated that the potential gains are worth exploring further because they offer an order of magnitude better performance.

So, I wanted to model performance to see if it was likely to result in a gain before doing a full implementation which might be a two or three month effort to get working well under all circumstances, documented, optimized for multi-threading etc. 

Steps in my process:

1) Model performance of iteration and creation
2) Share results and discuss
3) Model performance of indexed lookups and inbound set iteration

We're on step 2, and the next step will be to address this issue to see if this approach will also help inbound set searching. To do this will require implementing the on-page searching for index pages. In the case of the EdgePages, this will be used to quickly find the page which contains the inbound links for a given type, or type and atom. The way this works in a typical b-tree is to start at the root page, search that page for the sub-page containing the target type and atom, then search the sub-page, recursing until you get to a leaf page. 

From that point, since the leaf edge pages have a linear array layout, and these leaf pages are connected with pointers to adjacent pages, iterations over incoming sets will be linear traversals of memory which is the fastest iteration access pattern for hardware. Since the pages are relatively small, the shuffling around required to maintain sorted order for insertions and deletions is relatively minor, especially when compared with a large vector which might contain all the inbound atoms, for example. In the case of a 1K page, you'd be moving a few hundred bytes in a few already cached lines in the typical case. Only testing will show for certain.

Another approach worth testing and possible for the edge case is to add a vector with entries with the pointer to the leaf edge page for every type, that way one could simply look up the page in the vector and then search only that leaf page and not have to do a binary search. It may be that this approach could totally replace the b-tree itself for edge pages.

In the current prototype code, there are calls to memmove that are there to perform work that is approximately the same as what will be done in phase II when the actual search and sorting is implemented. What is not currently implemented is the actual comparison functions.

In the case of EdgePages where inbound links are stored, since each edge-pair is 16 bytes, and cache lines are 64 bytes, it may be that a straight linear search is faster than the more traditional binary search. For a 1K page, there are 16 64-byte cache lines, each containing 4 16-byte pairs. It may be that linear searches that check every fourth element at first and then check the skipped items on the first comparison that is greater than the target will be faster than a pure binary search because of the cache line pre-fetching hardware works well for linear searches at regular offsets in either ascending or descending order.

I expect this approach to exceed the performance of the current Map to Set standard containers, especially for larger incoming sets because there will be much less pointer chasing.

> How does this reconcile with the current AtomSpaceBenchmark results, on your machine? I'm certain that your machine is 2x or 3x faster than mine (mine is about 5-6 years old, now) and I'm getting addNode times of about 3μs and addLink times of 10μs so don't understand why you would be getting 25 μs on what must surely be a much faster machine.

I have been testing using a fairly old and slow machine from Ben's AI hedge-fund days which is a Dell PowerEdge T420 with a 6-core Xeon E5-2420 @ 1.90GHz and 32G of RAM. It is about 5 or 6 years old, as well.

Also, remember that the per-op time is for the EvaluationLink which is actually a creation of two link atoms, the EvaluationLink and the ListLink. There are no nodes created during the timed benchmarks. This corresponds to creating observations on word pairs, the predicate atom and the word nodes will already have been created so the additions will be the EvaluationLink and (perhaps) the ListLink for pairs that have not been observed before. In the benchmark, there is a pool of 100,000 nodes created in the test setup that are randomly paired for the ListLink, so most of the pairs will be new and result in a ListLink being created.

I reran some of these tests and am now getting numbers between 21 to 22 μs fairly consistently. This seems pretty close to your measurements, as it corresponds to a 10.5 to 11.0 μs per link.